agiles_service/requirements.txt
Flask
google-cloud-storage
python-dotenv
